refactor: Waiting 이름 ReservationWaiting으로 수정
