+++ src/main/java/roomescape/domain/ReservationWaiting.java
@@ -5,7 +5,7 @@ import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-public class Waiting {
+public class ReservationWaiting {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -27,14 +27,14 @@ public class Waiting {
     @JoinColumn(nullable = false)
     private Theme theme;
 
-    public Waiting() {
+    public ReservationWaiting() {
     }
 
-    public Waiting(Member member, ReservationDate date, ReservationTime time, Theme theme) {
+    public ReservationWaiting(Member member, ReservationDate date, ReservationTime time, Theme theme) {
         this(null, member, date, time, theme);
     }
 
-    public Waiting(Long id, Member member, ReservationDate date, ReservationTime time, Theme theme) {
+    public ReservationWaiting(Long id, Member member, ReservationDate date, ReservationTime time, Theme theme) {
         this.member = member;
         this.id = id;
         this.date = date;
@@ -42,12 +42,12 @@ public class Waiting {
         this.theme = theme;
     }
 
-    public static Waiting createIfFuture(LocalDateTime now, Member member, ReservationDate date, ReservationTime time, Theme theme) {
+    public static ReservationWaiting create(LocalDateTime now, Member member, ReservationDate date, ReservationTime time, Theme theme) {
         LocalDateTime waitingDateTime = LocalDateTime.of(date.getDate(), time.getStartAt());
         if (waitingDateTime.isBefore(now)) {
             throw new IllegalArgumentException(String.format("지나간 시간에 대한 에약 대기는 생성할 수 없습니다. (dateTime: %s)", waitingDateTime));
         }
-        return new Waiting(member, date, time, theme);
+        return new ReservationWaiting(member, date, time, theme);
     }
 
     public boolean hasSameMemberWith(Reservation reservation) {
