+++ src/main/java/roomescape/web/controller/response/ReservationWaitingWebResponse.java
@@ -1,0 +1,23 @@
+package roomescape.web.controller.response;
+
+import roomescape.service.response.ReservationWaitingAppResponse;
+
+import java.time.LocalDate;
+
+public record ReservationWaitingWebResponse(
+        Long id,
+        String name,
+        LocalDate date,
+        ReservationTimeWebResponse time,
+        ThemeWebResponse theme) {
+
+    public ReservationWaitingWebResponse(ReservationWaitingAppResponse response) {
+        this(
+                response.id(),
+                response.name(),
+                response.date().getDate(),
+                ReservationTimeWebResponse.from(response.reservationTimeAppResponse()),
+                ThemeWebResponse.from(response.themeAppResponse())
+        );
+    }
+}
