+++ src/test/java/roomescape/service/ReservationWaitingServiceTest.java
@@ -8,12 +8,18 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.*;
 import roomescape.infrastructure.*;
+import roomescape.service.request.WaitingAppRequest;
+import roomescape.service.response.ReservationWaitingAppResponse;
+import roomescape.service.response.ReservationWaitingWithRankAppResponse;
+import roomescape.web.exception.AuthorizationException;
 
 import java.time.LocalDate;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static roomescape.Fixture.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static roomescape.Fixture.VALID_RESERVATION_TIME;
+import static roomescape.Fixture.VALID_THEME;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @Sql(scripts = "/truncate.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
@@ -36,45 +42,173 @@ class ReservationWaitingServiceTest {
     private MemberRepository memberRepository;
 
     @Autowired
-    private WaitingRepository waitingRepository;
+    private ReservationWaitingRepository reservationWaitingRepository;
+
+    private final List<Member> members = List.of(
+            new Member(
+                    new MemberName("감자"),
+                    new MemberEmail("111@aaa.com"),
+                    new MemberPassword("asd"),
+                    MemberRole.USER
+            ),
+            new Member(
+                    new MemberName("고구마"),
+                    new MemberEmail("222@aaa.com"),
+                    new MemberPassword("asd"),
+                    MemberRole.USER
+            ),
+            new Member(
+                    new MemberName("단호박"),
+                    new MemberEmail("333@aaa.com"),
+                    new MemberPassword("asd"),
+                    MemberRole.USER
+            )
+    );
 
     @Test
-    @DisplayName("예약 대기가 존재하지 않는 경우 예약을 삭제한다.")
-    void deleteReservationIfNoWaiting() {
-        Member member = memberRepository.save(VALID_MEMBER);
+    @DisplayName("올바르게 예약 대기를 생성한다.")
+    void save() {
+        Member otherMember = memberRepository.save(members.get(0));
+        Member member = memberRepository.save(members.get(1));
+        String date = LocalDate.now().plusDays(1).toString();
         Theme theme = themeRepository.save(VALID_THEME);
         ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
-        String date = LocalDate.now().plusDays(2).toString();
-        Reservation reservation = reservationRepository.save(new Reservation(member, new ReservationDate(date), time, theme));
+        reservationRepository.save(new Reservation(otherMember, new ReservationDate(date), time, theme));
 
-        reservationWaitingService.deleteIfNoWaitingOrUpdateReservation(reservation.getId());
+        WaitingAppRequest request = new WaitingAppRequest(date, time.getId(), theme.getId(), member.getId());
+        ReservationWaitingAppResponse response = reservationWaitingService.save(request);
+        ReservationWaiting expectedWaiting = reservationWaitingRepository.findById(response.id()).get();
 
-        assertThat(reservationRepository.findById(reservation.getId())).isEmpty();
+        assertThat(response).isEqualTo(new ReservationWaitingAppResponse(expectedWaiting));
     }
 
     @Test
-    @DisplayName("예약 대기가 존재하는 경우, 우선순위가 높은 대기를 예약으로 변경한다.")
-    void updateReservation() {
-        Member reservedMember = memberRepository.save(VALID_MEMBER);
-        Member waitingMember = memberRepository.save(new Member(new MemberName("감자"),
-                new MemberEmail("111@aaa.com"),
-                VALID_USER_PASSWORD,
-                MemberRole.USER));
+    @DisplayName("본인이 예약한 날짜, 시간, 테마에 대해서 대기 생성 시 예외가 발생한다.")
+    void saveWithSelfReservationExist() {
+        Member member = memberRepository.save(members.get(0));
+        String date = LocalDate.now().plusDays(1).toString();
         Theme theme = themeRepository.save(VALID_THEME);
         ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
-        ReservationDate date = new ReservationDate(LocalDate.now().plusDays(2).toString());
-        Reservation reservation = reservationRepository.save(new Reservation(reservedMember, date, time, theme));
-        Waiting waiting = waitingRepository.save(new Waiting(waitingMember, date, time, theme));
+        reservationRepository.save(new Reservation(member, new ReservationDate(date), time, theme));
 
-        reservationWaitingService.deleteIfNoWaitingOrUpdateReservation(reservation.getId());
+        WaitingAppRequest request = new WaitingAppRequest(date, time.getId(), theme.getId(), member.getId());
 
-        Reservation updatedReservation = reservationRepository.findById(reservation.getId()).orElseThrow();
-        assertAll(
-                () -> assertThat(waitingRepository.findById(waiting.getId())).isEmpty(),
-                () -> assertThat(updatedReservation.getDate().getDate()).isEqualTo(date.getDate()),
-                () -> assertThat(updatedReservation.getMember()).isEqualTo(waiting.getMember()),
-                () -> assertThat(updatedReservation.getTheme().getId()).isEqualTo(theme.getId()),
-                () -> assertThat(updatedReservation.getTime().getId()).isEqualTo(time.getId())
-        );
+        assertThatThrownBy(() -> reservationWaitingService.save(request))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("본인이 예약한 날짜, 시간, 테마에 대해서는 대기를 생성할 수 없습니다.");
+    }
+
+    @Test
+    @DisplayName("예약이 없는 날짜, 시간, 테마에 대해서 대기 생성 시 예외가 발생한다.")
+    void saveWithNoReservationExist() {
+        Member member = memberRepository.save(members.get(0));
+        String date = LocalDate.now().plusDays(1).toString();
+        Theme theme = themeRepository.save(VALID_THEME);
+        ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
+
+        WaitingAppRequest request = new WaitingAppRequest(date, time.getId(), theme.getId(), member.getId());
+
+        assertThatThrownBy(() -> reservationWaitingService.save(request))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("예약이 존재하지 않는 날짜, 시간, 테마에 대해서는 대기를 생성할 수 없습니다.");
+    }
+
+    @Test
+    @DisplayName("날짜, 시간, 테마에 대해서 중복된 대기 생성 시 예외가 발생한다.")
+    void saveWithDuplicatedWaiting() {
+        Member otherMember = memberRepository.save(members.get(0));
+        Member member = memberRepository.save(members.get(1));
+        String date = LocalDate.now().plusDays(1).toString();
+        Theme theme = themeRepository.save(VALID_THEME);
+        ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
+        reservationRepository.save(new Reservation(otherMember, new ReservationDate(date), time, theme));
+        reservationWaitingRepository.save(new ReservationWaiting(member, new ReservationDate(date), time, theme));
+
+        WaitingAppRequest request = new WaitingAppRequest(date, time.getId(), theme.getId(), member.getId());
+
+        assertThatThrownBy(() -> reservationWaitingService.save(request))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("동일한 사용자의 중복된 예약 대기를 생성할 수 없습니다. ");
+    }
+
+    @Test
+    @DisplayName("지나간 시간에 대한 대기 생성 시 예외가 발생한다.")
+    void saveWithPastTime() {
+        Member otherMember = memberRepository.save(members.get(0));
+        Member member = memberRepository.save(members.get(1));
+        String date = LocalDate.now().minusDays(1).toString();
+        Theme theme = themeRepository.save(VALID_THEME);
+        ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
+        reservationRepository.save(new Reservation(otherMember, new ReservationDate(date), time, theme));
+
+        WaitingAppRequest request = new WaitingAppRequest(date, time.getId(), theme.getId(), member.getId());
+
+        assertThatThrownBy(() -> reservationWaitingService.save(request))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("지나간 시간에 대한 에약 대기는 생성할 수 없습니다.");
+    }
+
+    @Test
+    @DisplayName("지나간 시간에 대한 대기 생성 시 예외가 발생한다.")
+    void findWaitingWithRankByMemberId() {
+        List<Member> savedMembers = members.stream()
+                .map(memberRepository::save)
+                .toList();
+        String date = LocalDate.now().minusDays(1).toString();
+        Theme theme = themeRepository.save(VALID_THEME);
+        ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
+        reservationRepository.save(new Reservation(savedMembers.get(0), new ReservationDate(date), time, theme));
+        reservationWaitingRepository.save(new ReservationWaiting(savedMembers.get(1), new ReservationDate(date), time, theme));
+        ReservationWaiting savedWaiting = reservationWaitingRepository.save(new ReservationWaiting(savedMembers.get(2), new ReservationDate(date), time, theme));
+
+        List<ReservationWaitingWithRankAppResponse> waitingWithRankAppResponses = reservationWaitingService.findWaitingWithRankByMemberId(savedMembers.get(2).getId());
+
+        assertThat(waitingWithRankAppResponses)
+                .hasSize(1)
+                .containsExactly(new ReservationWaitingWithRankAppResponse(new ReservationWaitingWithRank(savedWaiting, 2L)));
+    }
+
+    @Test
+    @DisplayName("사용자의 예약 대기를 올바르게 삭제한다.")
+    void deleteMemberWaiting() {
+        Member reservedMember = memberRepository.save(members.get(0));
+        Member waitingMember = memberRepository.save(members.get(1));
+        String date = LocalDate.now().minusDays(1).toString();
+        Theme theme = themeRepository.save(VALID_THEME);
+        ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
+        reservationRepository.save(new Reservation(reservedMember, new ReservationDate(date), time, theme));
+        ReservationWaiting savedWaiting = reservationWaitingRepository.save(new ReservationWaiting(waitingMember, new ReservationDate(date), time, theme));
+
+        reservationWaitingService.deleteMemberWaiting(waitingMember.getId(), savedWaiting.getId());
+
+        assertThat(reservationWaitingRepository.findById(savedWaiting.getId())).isEmpty();
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 예약 대기 삭제 시 예외가 발생한다.")
+    void deleteNotExistWaiting() {
+        Member member = memberRepository.save(members.get(0));
+
+        assertThatThrownBy(() ->
+                reservationWaitingService.deleteMemberWaiting(member.getId(), 1L))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("예약 대기 삭제 실패: 대기를 찾을 수 없습니다.");
+    }
+
+    @Test
+    @DisplayName("자신이 생성하지 않은 예약 대기 삭제 시 예외가 발생한다.")
+    void deleteWaitingWithInvalidMember() {
+        Member reservedMember = memberRepository.save(members.get(0));
+        Member waitingMember = memberRepository.save(members.get(1));
+        String date = LocalDate.now().minusDays(1).toString();
+        Theme theme = themeRepository.save(VALID_THEME);
+        ReservationTime time = reservationTimeRepository.save(VALID_RESERVATION_TIME);
+        reservationRepository.save(new Reservation(reservedMember, new ReservationDate(date), time, theme));
+        ReservationWaiting savedWaiting = reservationWaitingRepository.save(new ReservationWaiting(waitingMember, new ReservationDate(date), time, theme));
+
+        assertThatThrownBy(() ->
+                reservationWaitingService.deleteMemberWaiting(reservedMember.getId(), savedWaiting.getId()))
+                .isInstanceOf(AuthorizationException.class)
+                .hasMessageContaining("예약 대기 삭제 권한이 없는 사용자입니다.");
     }
 }
